+++ src/cruds/core.py
@@ -3,9 +3,11 @@ Clients that can be used for easily accessing RESTful APIs
 """
 
 import abc
+from base64 import b64encode
 import logging
 from json.decoder import JSONDecodeError
 import sys
+from typing import Final
 
 from typing import Any, Dict, Union
 from urllib.parse import urlencode
@@ -13,12 +15,12 @@ from urllib.parse import urlencode
 import certifi
 import urllib3
 
-logger= logging.getLogger(__name__)
+logger = logging.getLogger(__name__)
 
-DEFAULT_TIMEOUT = 300.0
+DEFAULT_TIMEOUT: Final = 300.0
 
 
-class Auth(metaclass=abc.ABCMeta):
+class AuthABC(metaclass=abc.ABCMeta):
     """
     An Abstract Base Class that is used for implement the required interface
     that is used for Authentcation by the Client class.
@@ -54,7 +56,7 @@ class Client:
 
     Instance Attributes
     -------------------
-    status_ignore : set
+    status_ignore: set
         A set of status codes to ignore by instances with raise for status
         enabled.
 
@@ -70,18 +72,19 @@ class Client:
         Makes a DELETE request to the API Server
     """
 
-    def __init__(self,
-                 host: str,
-                 auth=None,
-                 manager=None,
-                 timeout=DEFAULT_TIMEOUT,
-                 raise_status=True,
-                 retries=4,
-                 backoff_factor=0.9,
-                 retry_status_codes=(504, 503, 502, 500, 429),
-                 serialize=True,
-                 verify_ssl=True,
-                 ) -> None:
+    def __init__(
+        self,
+        host: str,
+        auth=None,
+        manager=None,
+        timeout=DEFAULT_TIMEOUT,
+        raise_status=True,
+        retries=4,
+        backoff_factor=0.9,
+        retry_status_codes=(504, 503, 502, 500, 429, 425, 408),
+        serialize=True,
+        verify_ssl=True,
+    ) -> None:
         """
         Arguments
         ---------
@@ -90,7 +93,7 @@ class Client:
         auth: str, tuple, cruds.auth.Auth (optional)
             A bearer token can be supplied, or a tuple with the username
             and password. CRUDs includes more complex authentication using
-            the Auth Classes under the `cruds.auth` module.
+            the AuthABC Classes under the `cruds.auth` module.
         manager: urllib3.PoolManager, urllib3.ProxyManager (optional)
             You can supply a PoolManager with custom configuration.
         timeout: float (optional)
@@ -105,7 +108,7 @@ class Client:
             How much delay should be added with each retry.
         retry_status_codes: typle[int] (optional)
             Status codes that will trigger retries.
-            (default is (504, 503, 502, 429))
+            (default is (504, 503, 502, 500, 429, 425))
         serialize: boolaen (optional)
             Serialize and Deserialize dictionaires for data sent and received.
             (default is True)
@@ -115,62 +118,73 @@ class Client:
         """
         self.host: str = host if host.endswith("/") else host + "/"
         self.serialize: bool = serialize
-        logger.info("API Operation Timeout(sec): %s, Raises Exceptions on status: %s",
-                    timeout,
-                    raise_status)
+        logger.info(
+            "API Operation Timeout(sec): %s, Raises Exceptions on status: %s",
+            timeout,
+            raise_status,
+        )
 
         self.raise_status: bool = raise_status
         self.status_ignore: set = set()
-        self.timeout: float = timeout
 
         if isinstance(manager, (urllib3.PoolManager, urllib3.ProxyManager)):
             logger.info("Using supplied URLLib3 PoolManager or ProxyManager")
-            self._http = manager
+            self.manager = manager
         else:
             # Setup Retries
             if retries:
-                logger.info("Retries: %s attempts (backoff factor %s) for status codes %s",
-                            retries,
-                            backoff_factor,
-                            ', '.join([str(i) for i in retry_status_codes]))
+                logger.info(
+                    "Retries: %s attempts (backoff factor %s) for status codes %s",
+                    retries,
+                    backoff_factor,
+                    ", ".join([str(i) for i in retry_status_codes]),
+                )
 
-                retry = urllib3.Retry(connect=retries,
-                                      backoff_factor=backoff_factor,
-                                      status_forcelist=retry_status_codes)
+                retry = urllib3.Retry(
+                    total=retries,
+                    status_forcelist=retry_status_codes,
+                    backoff_factor=backoff_factor,
+                )
             else:
                 logger.info("Retries: Disabled")
                 retry = False
 
             # Create PoolManager
-            self._http = urllib3.PoolManager(
+            self.manager = urllib3.PoolManager(
                 cert_reqs="CERT_REQUIRED" if verify_ssl else "CERT_NONE",
                 ca_certs=certifi.where(),
-                retries=retry)
+                retries=retry,
+                timeout=timeout,
+            )
 
         # Setup Headers (Authentication)
-        self._request_headers = {}
+        self.request_headers = urllib3.HTTPHeaderDict()
 
         if serialize is True:
-            self._request_headers["Content-Type"]  = "application/json"
+            self.request_headers.add("Content-Type", "application/json")
 
         if isinstance(auth, str):
-            self._request_headers["Authorization"] = f"Bearer {auth}"
+            self.request_headers.add("Authorization", f"Bearer {auth}")
             logger.info("Token authentication setup")
         elif isinstance(auth, (list, tuple)) and len(auth) == 2:
-            self._request_headers = urllib3.make_headers(basic_auth=":".join(auth))
+            self.request_headers.add(
+                "Authorization",
+                "Basic " + b64encode(":".join(auth).encode("UTF-8")).decode("UTF-8"),
+            )
             logger.info("Basic authentication setup")
-        elif isinstance(auth, Auth):
-            self.auth: Auth = auth
+        elif isinstance(auth, AuthABC):
+            self.auth: AuthABC = auth
             self._check_auth()
             logger.info(f"{auth.__class__.__name__} authentication setup")
         else:
             logger.info("No authentication setup")
 
-    def create(self,
-               uri: str,
-               data: dict,
-               params: Union[Dict[Any, Any], None] = None,
-               ) -> Union[Dict[Any, Any], bytes]:
+    def create(
+        self,
+        uri: str,
+        data: dict,
+        params: Union[Dict[Any, Any], None] = None,
+    ) -> Union[Dict[Any, Any], bytes]:
         """
         Makes a basic Create request to the API, and returns the response.
 
@@ -201,24 +215,21 @@ class Client:
 
         self._check_auth()
         if self.serialize and isinstance(data, dict):
-            response = self._http.request(method,
-                                          url + safe_params,
-                                          headers=self._request_headers,
-                                          json=data,
-                                          timeout=self.timeout)
+            response = self.manager.request(
+                method, url + safe_params, headers=self.request_headers, json=data
+            )
         else:
-            response = self._http.request(method,
-                                          url + safe_params,
-                                          body=data,
-                                          headers=self._request_headers,
-                                          timeout=self.timeout)
+            response = self.manager.request(
+                method, url + safe_params, body=data, headers=self.request_headers
+            )
 
         return self._process_resp(method, response)
 
-    def read(self,
-             uri: str,
-             params: Union[Dict[Any, Any], None] = None,
-             ) -> Union[Dict[Any, Any], bytes]:
+    def read(
+        self,
+        uri: str,
+        params: Union[Dict[Any, Any], None] = None,
+    ) -> Union[Dict[Any, Any], bytes]:
         """
         Makes a basic Retrieve request to the API, and returns the response
 
@@ -240,19 +251,18 @@ class Client:
         logger.info(f"API Retrieve Operation to {url}")
 
         self._check_auth()
-        response = self._http.request(method,
-                                      url,
-                                      fields=params,
-                                      headers=self._request_headers,
-                                      timeout=self.timeout)
+        response = self.manager.request(
+            method, url, fields=params, headers=self.request_headers
+        )
         return self._process_resp(method, response)
 
-    def update(self,
-               uri: str,
-               data: Union[Dict[Any, Any], str],
-               params: Union[Dict[Any, Any], None] = None,
-               replace: bool = False,
-               ) -> Union[Dict[Any, Any], bytes]:
+    def update(
+        self,
+        uri: str,
+        data: Union[Dict[Any, Any], str],
+        params: Union[Dict[Any, Any], None] = None,
+        replace: bool = False,
+    ) -> Union[Dict[Any, Any], bytes]:
         """
         Makes a basic Update request to the API, and returns the response.
 
@@ -285,24 +295,19 @@ class Client:
 
         self._check_auth()
         if self.serialize and isinstance(data, dict):
-            response = self._http.request(method,
-                                          url + safe_params,
-                                          headers=self._request_headers,
-                                          json=data,
-                                          timeout=self.timeout)
+            response = self.manager.request(
+                method, url + safe_params, headers=self.request_headers, json=data
+            )
         else:
-            response = self._http.request(method,
-                                          url + safe_params,
-                                          body=data,
-                                          headers=self._request_headers,
-                                          timeout=self.timeout)
+            response = self.manager.request(
+                method, url + safe_params, body=data, headers=self.request_headers
+            )
 
         return self._process_resp(method, response)
 
-    def delete(self,
-               uri: str,
-               params: Union[Dict[Any, Any], None] = None
-               ) -> Union[Dict[Any, Any], bytes]:
+    def delete(
+        self, uri: str, params: Union[Dict[Any, Any], None] = None
+    ) -> Union[Dict[Any, Any], bytes]:
         """
         Makes a basic Delete request to the API, and returns the response
 
@@ -322,18 +327,16 @@ class Client:
         logger.info(f"API Delete Operation to {url}")
 
         self._check_auth()
-        response = self._http.request(method,
-                                      self.host + uri,
-                                      fields=params,
-                                      headers=self._request_headers,
-                                      timeout=self.timeout)
+        response = self.manager.request(
+            method, self.host + uri, fields=params, headers=self.request_headers
+        )
         return self._process_resp(method, response)
 
     def _process_resp(
-            self,
-            method: str,
-            response: urllib3.response.BaseHTTPResponse,
-            ) -> Union[Dict[Any, Any], bytes]:
+        self,
+        method: str,
+        response: urllib3.response.BaseHTTPResponse,
+    ) -> Union[Dict[Any, Any], bytes]:
         """
         Processes the Responce from URLLib3 request in a standardize manner, and
         displays information.
@@ -352,15 +355,19 @@ class Client:
                 error_type = None
 
             if error_type:
-                msg: str = f"{error_type} Error with status code {response.status}" \
-                      f" Message: {response.data.decode('utf-8')}"
+                msg: str = (
+                    f"{error_type} Error with status code {response.status}"
+                    f" Message: {response.data.decode('utf-8')}"
+                )
                 raise urllib3.exceptions.HTTPError(msg)
 
         if self.serialize:
-            if 'application/json' in response.headers.get('Content-Type', ''):
+            if "application/json" in response.headers.get("Content-Type", ""):
                 return response.json()
 
-            logger.warning("Response content type is not declared as JSON but serialize is enabled")
+            logger.warning(
+                "Response content type is not declared as JSON but serialize is enabled"
+            )
             try:
                 return response.json()
             except JSONDecodeError:
@@ -369,5 +376,11 @@ class Client:
         return response.data
 
     def _check_auth(self):
-        if hasattr(self, "auth") and not self.auth.is_valid():
-            self._request_headers["Authorization"] = "Bearer " + self.auth.access_token()
+        if (
+            hasattr(self, "auth")
+            and isinstance(self.auth, AuthABC)
+            and not self.auth.is_valid()
+        ):
+            self.request_headers.add(
+                "Authorization", "Bearer " + self.auth.access_token()
+            )