+++ src/cruds/exception.py
@@ -2,5 +2,6 @@
 Custom Exceptions
 """
 
+
 class OAuthAccessTokenError(Exception):
     pass