+++ src/cruds/interfaces/planhat/exception.py
@@ -2,5 +2,6 @@
 Custom Exceptions specifically for an Interface
 """
 
+
 class PlanhatUpsertError(Exception):
     pass