+++ tests/test_auth.py
@@ -1,6 +1,7 @@
 """
 Tests for Auth components in CRUDs
 """
+
 import json
 from time import time
 from unittest import mock
@@ -12,20 +13,24 @@ from cruds.auth import OAuth2
 access_token: str = "MTQ0NjJkZmQ5OTM2NDE1ZTZjNGZmZjI3"
 refresh_token: str = "IwOGYzYTlmM2YxOTQ5MGE3YmNmMDFkNTVk"
 
-access_token_response: bytes = json.dumps({
-  "access_token": access_token,
-  "token_type": "Bearer",
-  "expires_in": 60,
-  "scope": "api",
-}).encode("utf-8")
+access_token_response: bytes = json.dumps(
+    {
+        "access_token": access_token,
+        "token_type": "Bearer",
+        "expires_in": 60,
+        "scope": "api",
+    }
+).encode("utf-8")
 
-access_token_response_with_refresh_token: bytes = json.dumps({
-  "access_token": access_token,
-  "token_type": "Bearer",
-  "expires_in": 60,
-  "refresh_token": refresh_token,
-  "scope": "api",
-}).encode("utf-8")
+access_token_response_with_refresh_token: bytes = json.dumps(
+    {
+        "access_token": access_token,
+        "token_type": "Bearer",
+        "expires_in": 60,
+        "refresh_token": refresh_token,
+        "scope": "api",
+    }
+).encode("utf-8")
 
 
 def test_OAuth2_empty_state():
@@ -71,7 +76,6 @@ def test_OAuth2_access_token_cached():
     mock_request.assert_not_called()
 
 
-
 def test_OAuth2_access_token_request():
     """
     If the token has not been cached in the Class state, then the client credientials
@@ -101,7 +105,7 @@ def test_OAuth2_access_token_request():
         headers={
             "authorization": "Basic MTIzOkFCQw==",
             "Content-Type": "application/x-www-form-urlencoded; charset=utf-8",
-            },
+        },
         redirect=False,
     )
 
@@ -138,7 +142,7 @@ def test_OAuth2_access_token_request_with_password_grant_type():
         headers={
             "authorization": "Basic MTIzOkFCQw==",
             "Content-Type": "application/x-www-form-urlencoded; charset=utf-8",
-            },
+        },
         redirect=False,
     )
 
@@ -177,6 +181,6 @@ def test_OAuth2_access_token_refresh():
         body=f"grant_type=refresh_token&refresh_token={refresh_token}",
         headers={
             "Content-Type": "application/x-www-form-urlencoded; charset=utf-8",
-            },
+        },
         redirect=False,
     )