+++ tests/test_core.py
@@ -8,11 +8,10 @@ import pytest
 import urllib3
 
 import cruds
-from cruds.core import Auth, DEFAULT_TIMEOUT
+from cruds.core import AuthABC
 
-request_headers = {
-    "Content-Type": "application/json"
-}
+request_headers = urllib3.HTTPHeaderDict({"Content-Type": "application/json"})
+
 
 def test_Client_token_authentication():
     """
@@ -20,7 +19,7 @@ def test_Client_token_authentication():
     authenticationi.
     """
     api = cruds.Client(host="https://localhost", auth="api_token")
-    assert api._request_headers.get("Authorization") == "Bearer api_token"
+    assert api.request_headers.get("Authorization") == "Bearer api_token"
 
 
 def test_Client_basic_authentication():
@@ -29,7 +28,7 @@ def test_Client_basic_authentication():
     authentication.
     """
     api = cruds.Client(host="https://localhost", auth=("username", "password"))
-    assert api._request_headers.get("authorization") == "Basic dXNlcm5hbWU6cGFzc3dvcmQ="
+    assert api.request_headers.get("authorization") == "Basic dXNlcm5hbWU6cGFzc3dvcmQ="
 
 
 def test_Client_creates_urllib3_manager():
@@ -37,7 +36,7 @@ def test_Client_creates_urllib3_manager():
     When no URLLib3 manager is suppied, a manager is automatically created.
     """
     api = cruds.Client(host="https://localhost")
-    assert isinstance(api._http, urllib3.PoolManager)
+    assert isinstance(api.manager, urllib3.PoolManager)
 
 
 def test_Client_use_supplied_urllib3_manager():
@@ -45,16 +44,16 @@ def test_Client_use_supplied_urllib3_manager():
     When a URLLib3 manager is suppied use it, instead of creating one.
     """
     api = cruds.Client(host="https://localhost", manager=urllib3.PoolManager())
-    assert isinstance(api._http, urllib3.PoolManager)
+    assert isinstance(api.manager, urllib3.PoolManager)
 
 
 def test_Client_disable_retries():
-    """ Setting the retries to 0 or None will disable retries being used """
+    """Setting the retries to 0 or None will disable retries being used"""
     api = cruds.Client(host="https://localhost", retries=0)
-    assert api._http.connection_pool_kw.get("retries") == 0
+    assert api.manager.connection_pool_kw.get("retries") == 0
 
     cruds.Client(host="https://localhost", retries=None)
-    assert api._http.connection_pool_kw.get("retries") == 0
+    assert api.manager.connection_pool_kw.get("retries") == 0
 
 
 @pytest.fixture
@@ -64,21 +63,22 @@ def crud_api():
     """
     api = cruds.Client(host="https://localhost", retries=0)
     mock_resp = urllib3.HTTPResponse(body=b'{"name": "test"}')
-    api._http.request = mock.Mock(return_value=mock_resp)
+    api.manager.request = mock.Mock(return_value=mock_resp)
     api._process_resp = lambda method, resp: resp
     return api
 
 
 def test_Client_create_operation(crud_api):
-    """ Check the Create Operation formats the request properly """
+    """Check the Create Operation formats the request properly"""
     sample = {"test_name": "test_Client_create_operation"}
     resp = crud_api.create("user/1", data=sample)
 
-    crud_api._http.request.assert_called_with("POST",
-                                              "https://localhost/user/1",
-                                              headers=request_headers,
-                                              json=sample,
-                                              timeout=DEFAULT_TIMEOUT)
+    crud_api.manager.request.assert_called_with(
+        "POST",
+        "https://localhost/user/1",
+        headers=request_headers,
+        json=sample,
+    )
     assert resp.data == b'{"name": "test"}'
 
 
@@ -89,11 +89,12 @@ def test_Client_create_operation_with_bytes(crud_api):
     sample = b'{"test_name": "test_Client_create_operation"}'
     resp = crud_api.create("user/2", data=sample)
 
-    crud_api._http.request.assert_called_with("POST",
-                                              "https://localhost/user/2",
-                                              body=sample,
-                                              headers=request_headers,
-                                              timeout=DEFAULT_TIMEOUT)
+    crud_api.manager.request.assert_called_with(
+        "POST",
+        "https://localhost/user/2",
+        body=sample,
+        headers=request_headers,
+    )
     assert resp.data == b'{"name": "test"}'
 
 
@@ -103,11 +104,12 @@ def test_Client_read_operation(crud_api):
     """
     resp = crud_api.read("test")
 
-    crud_api._http.request.assert_called_with("GET",
-                                              "https://localhost/test",
-                                              fields=None,
-                                              headers=request_headers,
-                                              timeout=DEFAULT_TIMEOUT)
+    crud_api.manager.request.assert_called_with(
+        "GET",
+        "https://localhost/test",
+        fields=None,
+        headers=request_headers,
+    )
     assert resp.data == b'{"name": "test"}'
 
 
@@ -118,11 +120,12 @@ def test_Client_update_operation(crud_api):
     sample = {"test_name": "test_Client_update_operation"}
     resp = crud_api.update("test", data=sample)
 
-    crud_api._http.request.assert_called_with("PATCH",
-                                              "https://localhost/test",
-                                              headers=request_headers,
-                                              json=sample,
-                                              timeout=DEFAULT_TIMEOUT)
+    crud_api.manager.request.assert_called_with(
+        "PATCH",
+        "https://localhost/test",
+        headers=request_headers,
+        json=sample,
+    )
     assert resp.data == b'{"name": "test"}'
 
 
@@ -133,11 +136,12 @@ def test_Client_update_operation_with_bytes(crud_api):
     sample = b'{"test_name": "test_Client_update_operation"}'
     resp = crud_api.update("test", data=sample)
 
-    crud_api._http.request.assert_called_with("PATCH",
-                                              "https://localhost/test",
-                                              body=sample,
-                                              headers=request_headers,
-                                              timeout=DEFAULT_TIMEOUT)
+    crud_api.manager.request.assert_called_with(
+        "PATCH",
+        "https://localhost/test",
+        body=sample,
+        headers=request_headers,
+    )
     assert resp.data == b'{"name": "test"}'
 
 
@@ -148,11 +152,12 @@ def test_Client_update_operation_with_replace(crud_api):
     sample = {"test_name": "test_Client_update_operation"}
     resp = crud_api.update("test", data=sample, replace=True)
 
-    crud_api._http.request.assert_called_with("PUT",
-                                              "https://localhost/test",
-                                              headers=request_headers,
-                                              json=sample,
-                                              timeout=DEFAULT_TIMEOUT)
+    crud_api.manager.request.assert_called_with(
+        "PUT",
+        "https://localhost/test",
+        headers=request_headers,
+        json=sample,
+    )
     assert resp.data == b'{"name": "test"}'
 
 
@@ -162,11 +167,12 @@ def test_Client_delete_operation(crud_api):
     """
     resp = crud_api.delete("test")
 
-    crud_api._http.request.assert_called_with("DELETE",
-                                              "https://localhost/test",
-                                              fields=None,
-                                              headers=request_headers,
-                                              timeout=DEFAULT_TIMEOUT)
+    crud_api.manager.request.assert_called_with(
+        "DELETE",
+        "https://localhost/test",
+        fields=None,
+        headers=request_headers,
+    )
     assert resp.data == b'{"name": "test"}'
 
 
@@ -175,8 +181,13 @@ def test_Client_process_resp_return_bytes():
     Check the response processing returns bytes for non-JSON content.
     """
     api = cruds.Client(host="https://localhost")
-    mock_resp = urllib3.HTTPResponse(body=b"name=test_Client_process_resp_return_bytes", status=399)
-    assert api._process_resp("", mock_resp) == b"name=test_Client_process_resp_return_bytes"
+    mock_resp = urllib3.HTTPResponse(
+        body=b"name=test_Client_process_resp_return_bytes", status=399
+    )
+    assert (
+        api._process_resp("", mock_resp)
+        == b"name=test_Client_process_resp_return_bytes"
+    )
 
 
 def test_Client_process_resp_return_bytes_with_serialize_false():
@@ -186,9 +197,13 @@ def test_Client_process_resp_return_bytes_with_serialize_false():
     """
     api = cruds.Client(host="https://localhost", serialize=False)
     mock_resp = urllib3.HTTPResponse(
-            body=b'{"name": "test_Client_process_resp_return_bytes_with_serialize_false"}',
-            headers={"Content-Type": "application/json; charset=utf-8"})
-    assert api._process_resp("", mock_resp) == b'{"name": "test_Client_process_resp_return_bytes_with_serialize_false"}'
+        body=b'{"name": "test_Client_process_resp_return_bytes_with_serialize_false"}',
+        headers={"Content-Type": "application/json; charset=utf-8"},
+    )
+    assert (
+        api._process_resp("", mock_resp)
+        == b'{"name": "test_Client_process_resp_return_bytes_with_serialize_false"}'
+    )
 
 
 def test_Client_process_resp_return_dictionary():
@@ -197,10 +212,13 @@ def test_Client_process_resp_return_dictionary():
     """
     api = cruds.Client(host="https://localhost")
     mock_resp = urllib3.HTTPResponse(
-            body=b'{"name": "test_Client_process_resp_return_dictionary"}',
-            headers={"Content-Type": "application/json; charset=utf-8"},
-            status=399)
-    assert api._process_resp("", mock_resp) == {"name": "test_Client_process_resp_return_dictionary"}
+        body=b'{"name": "test_Client_process_resp_return_dictionary"}',
+        headers={"Content-Type": "application/json; charset=utf-8"},
+        status=399,
+    )
+    assert api._process_resp("", mock_resp) == {
+        "name": "test_Client_process_resp_return_dictionary"
+    }
 
 
 def test_Client_raise_status_399():
@@ -248,14 +266,15 @@ def test__check_auth_invalid():
     the request headers.
     """
 
-    class MockAuth(Auth):
+    class MockAuth(AuthABC):
         def access_token(self):
             return "9a8sdftg"
+
         def is_valid(self):
             return False
 
     api = cruds.Client(host="https://localhost", auth=MockAuth())
-    assert api._request_headers.get("Authorization") == "Bearer 9a8sdftg"
+    assert api.request_headers.get("Authorization") == "Bearer 9a8sdftg"
 
 
 def test__check_auth_still_valid():
@@ -263,11 +282,13 @@ def test__check_auth_still_valid():
     If the auth attribute is found and token is valid do nothing to the request
     headers.
     """
-    class MockAuth(Auth):
+
+    class MockAuth(AuthABC):
         def access_token(self):
-            return "9a8sdftg"
+            return "9a8sdftg"  # TODO Mock this to test side effects rather than None.
+
         def is_valid(self):
             return True
 
     api = cruds.Client(host="https://localhost", auth=MockAuth())
-    assert api._request_headers.get("Authorization") == None
+    assert api.request_headers.get("Authorization") == None