+++ docs/changelog.rst
@@ -1,21 +1,34 @@
 Changelog
 =========
 
+Release 1.3.1 (October 16, 2024)
+--------------------------------
+
+Updates:
+ - Renamed the ``Auth`` Abstract Base Class to ``AuthABC``.
+ - Made the PoolManager or ProxyManger attribute ``manager`` public.
+ - Made the HTTPHeaderDict attribute ``request_headers`` public.
+ - Retries applies to all URLLib3 retry types instead of only connect.
+ - Added 'Request Timeout' and 'Too Early' default retry status codes.
+ - status_whitelist renamed to status_ignore for better representation.
+
 Release 1.3.0 (October 7, 2024)
 -------------------------------
 
 Features:
- * OAuth 2.0 Rich Authorization Requests (`RFC 9396 <https://datatracker.ietf.org/doc/html/rfc9396>`_)
-
+ - OAuth 2.0 Rich Authorization Requests (`RFC 9396 <https://datatracker.ietf.org/doc/html/rfc9396>`_)
 
 Release 1.2.0 (August 7, 2024)
 ------------------------------
 
 Features:
- * Grant Types: Client Credentials and Password.
- * Tokens will refresh if the response from the server includes a Expires in field.
- * Refresh tokens will be used if included in the response from the server.
-
+ - Initial implmentation of OAuth 2.0 Authorization Grant
+     - Resource Owner Password Credientals
+       (`Section 1.3.3 RFC 6749 <https://www.rfc-editor.org/rfc/rfc6749#section-1.3.3>`_)
+     - Client Credentials
+       (`Section 1.3.4 RFC 6749 <https://www.rfc-editor.org/rfc/rfc6749#section-1.3.4>`_)
+     - Refresh tokens will be used if included in the response from the server.
+       (`Section 1.5 RFC 6749 <https://www.rfc-editor.org/rfc/rfc6749#section-1.5>`_)
 
 Release 1.1.0 (May 14, 2024)
 ----------------------------